Add files via upload
TinkerCAD task

The circuit has been given.
Link to circuit:https://www.tinkercad.com/things/lMwjnJX7suc?sharecode=RylPCnp5oT5dlws194qyk78kYNVbjncXosr-r17EK9E

Code using Arduino such that when the motor is running, LED is on and when the motor is not running, LED is off. The motor should run with a 5s delay and in a loop.